Refatoração das camadas para padronização

+++ teart/src/main/java/com/duckers/teart/repositories/Interfaces/EnderecoAtelierJPA.java
@@ -3,5 +3,5 @@ package com.duckers.teart.repositories.Interfaces;
 import com.duckers.teart.entities.EnderecoAtelier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoAtelierJPA extends JpaRepository<EnderecoAtelier, Integer> {
+public interface EnderecoAtelierJPA extends JpaRepository<EnderecoAtelier, Long> {
 }
